Moved MinesweeperMove to the `minesweeper` package, outside of interfaces
